Don't reset buddy list when moving it
